requirements.txt
python-telegram-bot>=20.5
spotdl>=4.5.0
youtube-dl>=2023.12.17

